Revert "Added Stop function to nodes as well as Digital XInput controls"
This reverts commit 590ceca032a73753429f9f68972dd88580a84f85.

+++ NekoPuppet/FunctionGraphForm.cs
@@ -55,15 +55,6 @@ namespace NekoPuppet
             });*/
         }
 
-        public void NotifyNodesOfClose()
-        {
-            List<FunctionalNetworkModel.NodeViewModel> nodes = this.nodeGraph1.ViewModel.Network.Nodes.ToList();
-            nodes.ForEach(node =>
-            {
-                //node = null; // will the GC be enough? NOPE
-                node.Stop();
-            });
-        }
 
         private IControlNodeFactoryPlugin[] LoadAssembly(string assemblyPath)
         {
@@ -97,7 +88,6 @@ namespace NekoPuppet
 
         private void NodeGraph1_NodeDeleted(object sender, NodeGraph.NodeDeletedEventArgs e)
         {
-            //e.Node.Stop();
             e.Node.Dispose();
         }
 
@@ -128,7 +118,7 @@ namespace NekoPuppet
                 var nodesCopy = this.nodeGraph1.ViewModel.Network.Nodes.ToArray();
                 foreach (var node in nodesCopy)
                 {
-                    //if (node.IsSelected)
+                    if (node.IsSelected)
                     {
                         nodeGraph1.ViewModel.DeleteNode(node);
                         node.Dispose();
@@ -162,7 +152,7 @@ namespace NekoPuppet
 
                                         model.X = dr["x"] != null && dr["x"].Type == JTokenType.Float ? (float)dr["x"] : 0f;
                                         model.Y = dr["y"] != null && dr["y"].Type == JTokenType.Float ? (float)dr["y"] : 0f;
-                                        NodeViewModel created = nodeGraph1.ViewModel.CreateNode(model, /*position,*/ false);
+                                        nodeGraph1.ViewModel.CreateNode(model, /*position,*/ true);
                                     }
                                 }
                             }
@@ -252,7 +242,7 @@ namespace NekoPuppet
 
                                     model.X = dr["x"] != null && dr["x"].Type == JTokenType.Float ? (float)dr["x"] : 0f;
                                     model.Y = dr["y"] != null && dr["y"].Type == JTokenType.Float ? (float)dr["y"] : 0f;
-                                    NodeViewModel created = nodeGraph1.ViewModel.CreateNode(model, /*position,*/ false);
+                                    nodeGraph1.ViewModel.CreateNode(model, /*position,*/ true);
                                 }
                             }
                         }
